unit tests framework: added global and batch timers.

+++ tests/maketests.sh
@@ -19,6 +19,7 @@ BUILDDIR="${BUILDIR:-${TESTS_DIR}/../build}"  # BME build directory ("compiled" 
 # It can get a list of tests to run as parameter
 function main() {
 local test_counter=0
+local test_start=$(date +%s)
 
 # list of tests to run
 	if (( ${#} > 0 )); then
@@ -49,6 +50,7 @@ local test_counter=0
 		local test_scratch_dir="${SCRATCH_BASE_DIR}/test_${padded_random}"
 		[ ${DEBUG:+1} ] && test_log "TEST's scratch dir: '${test_scratch_dir}'"
 		mkdir --parents ${test_scratch_dir}
+		local batch_start=$(date +%s)
 		env --ignore-environment \
 			PATH="$(realpath "${BUILDDIR}"):/usr/local/sbin:/usr/bin:/usr/sbin:/bin:/sbin" \
 			HOME="${test_scratch_dir}" \
@@ -56,23 +58,38 @@ local test_counter=0
 			bash -c "source "${TESTS_DIR}/helper_functions.sh" && ${test}"
 	# Now, check result from command above
 		local test_rc=$?
+		local batch_duration=$( seconds_duration $(( $(date +%s)-batch_start )) )
+		local batch_msg="${C_BOLD}${test_counter}. '${test}'${C_NC}\n"
+		batch_msg+="\tbatch time: ${T_BOLD}${batch_duration}${T_NC}\n"
 		if [ $test_rc -ne 0 ]; then
-			local err_msg="${C_BOLD}'${test}'${C_NC} (${test_rc})"
-			err_msg+="\n\tSee both the output above and the contents of the test's scratch dir:"
-			err_msg+="\n\t\t'${T_BOLD}${test_scratch_dir}'${T_NC}\n"
-			err_msg+="\n${C_BOLD}UNITARY TEST BATCHES RUN: ${C_YELLOW}${test_counter}${C_NC}."
+			batch_msg+="\t(${T_RED}${test_rc}${T_NC}) "
+			batch_msg+="See both the output above and the contents of the test's scratch dir:"
+			batch_msg+="\n\t\t'${T_BOLD}${test_scratch_dir}'${T_NC}\n"
 			echo ''
-			test_log "${err_msg}" error 0
-			exit $test_rc
+			test_log "${batch_msg}" error 0
+			break
 		fi
 	# Test finished OK; let's clean
 		rm --recursive --force ${test_scratch_dir}
+		test_log "${batch_msg}" ok 0
 	done
 	shopt -u nullglob globstar
-	# Finally, delete the "main" scratch dir
-	rm --recursive --force ${test_scratch_dir}
+
 	echo ''
-	test_log "${C_BOLD}UNITARY TEST BATCHES RUN: ${C_GREEN}${test_counter}${C_NC}." info 0
+	local final_msg="${T_BOLD}TEST BATCHES RUN:${T_NC} "
+	if (( ${test_rc} == 0 )); then
+		local final_status='info'
+		rm --recursive --force ${test_scratch_dir}
+		final_msg+="${T_GREEN}${test_counter}${T_NC}\n"
+	else
+		local final_status='fail'
+		final_msg+="${T_YELLOW}${test_counter}${T_NC}\n"
+	fi
+
+	local elapsed_time=$( seconds_duration $(( $(date +%s)-test_start )) )
+	final_msg+="\telapsed time: ${T_BOLD}${elapsed_time}${T_NC}"
+	test_log "${final_msg}" "${final_status}" 0
+	return ${test_rc}
 }
